Revert "Manage time in a better way"
