+++ sconstruct.py
@@ -94,7 +94,7 @@ smek_source.remove(smek_dir + "platform.cpp")  # The platform layer
 platform_source = [re.sub("^src/", smek_dir, f) for f in source if "imgui" in f or "glad" in f or "platform" in f]
 smek = env.Program(target=smek_dir + "SMEK", source=[platform_source, smek_dir + "util/tprint.cpp"])
 libsmek = env.SharedLibrary(target=smek_dir + "libSMEK", source=smek_source)
-AddPostAction(libsmek, "(pidof SMEK >/dev/null && kill -USR1 $$(pidof SMEK)) || true")
+AddPostAction(libsmek, "(pidof SMEK >/dev/null && kill -USR1 $$(pidof SMEK)) || true")is:pr is:open 
 Depends(smek, libsmek)
 Depends(smek, assets)
 AddPostAction(assets, "(pidof SMEK >/dev/null && kill -USR1 $$(pidof SMEK)) || true")
